+++ src/date-picker/hooks/useRangeValue.ts
@@ -23,6 +23,7 @@ export default function useRangeValue(props: TdDateRangePickerProps) {
     getDefaultFormat({
       mode: props.mode,
       format: props.format,
+      valueType: props.valueType,
       enableTimePicker: props.enableTimePicker,
     }),
   );
@@ -70,7 +71,8 @@ export default function useRangeValue(props: TdDateRangePickerProps) {
     if (!isValidDate(value.value, formatRef.value.format)) return;
 
     cacheValue.value = formatDate(value.value, {
-      format: formatRef.value.format,
+      format: formatRef.value.valueType,
+      targetFormat: formatRef.value.format,
     });
     time.value = formatTime(
       value.value,