+++ src/form/hooks.ts
@@ -10,8 +10,7 @@ export interface FormReadonlyProvider {
 }
 
 /**
- * 用于实现 form 的全局禁用状态hook
- * @returns
+ * @deprecated use src/hooks/useDisabled instead
  */
 export function useFormDisabled(extend?: Ref<boolean>) {
   const ctx = getCurrentInstance();