+++ src/table/_example-ts/editable-cell.vue
@@ -152,7 +152,7 @@ const columns = computed<TableProps['columns']>(() => [
     title: '申请状态',
     colKey: 'status',
     cell: (h, { row }) => STATUS_OPTIONS.find((t) => t.value === row.status)?.label,
-    1: {
+    edit: {
       component: Select,
       // props, 透传全部属性到 Select 组件
       props: {